made darkplaces able to compile as C++ again, and fixed all conversion warnings when doing so changed several malloc and free calls to Z_Malloc and Z_Free
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6260 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_backend.c
@@ -379,7 +379,7 @@ typedef struct gltextureunit_s
 	int arrayenabled;
 	unsigned int arraycomponents;
 	const void *pointer_texcoord;
-	float rgbscale, alphascale;
+	int rgbscale, alphascale;
 	int combinergb, combinealpha;
 	// FIXME: add more combine stuff
 	// texmatrixenabled exists only to avoid unnecessary texmatrix compares
